Move email_helpers to services. Optimise deployment strategy.
